# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Moldova Visa Assist Recruitment API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: number;
  title: string;
  /** e.g. IT, Finance, Healthcare, Construction, Hospitality */
  category: string;
  /** Country or city where the job is located */
  location: string;
  /** Full-time, Part-time, Contract */
  type: string;
  description: string;
  requirements: string;
  /** Salary range or negotiable */
  salary: string;
  /** @nullable */
  benefits?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface JobInput {
  title: string;
  category: string;
  location: string;
  type: string;
  description: string;
  requirements: string;
  salary: string;
  benefits?: string;
}

export interface Application {
  id: number;
  jobId: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  nationality?: string | null;
  dateOfBirth?: string | null;
  passportNumber?: string | null;
  yearsExperience?: string | null;
  skills?: string | null;
  languages?: string | null;
  availableFrom?: string | null;
  resumeUrl?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  experience?: string | null;
  status: string;
  createdAt: string;
}

export interface ApplicationInput {
  jobId: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  nationality?: string;
  dateOfBirth?: string;
  passportNumber?: string;
  yearsExperience?: string;
  skills?: string;
  languages?: string;
  availableFrom?: string;
  resumeUrl?: string;
  coverLetter?: string;
  experience?: string;
}

export interface WorkPermitInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  nationality: string;
  dateOfBirth: string;
  passportNumber: string;
  passportExpiry: string;
  currentAddress: string;
  permitType: string;
  employerName: string;
  employerCountry: string;
  jobTitle: string;
  jobSalary: string;
  startDate: string;
  contractDuration: string;
  hasPassport: boolean;
  hasJobOffer: boolean;
  hasMedicalCert: boolean;
  hasCriminalRecord: boolean;
  hasPhotos: boolean;
  hasEducationCert: boolean;
}

export interface WorkPermit {
  id: number;
  referenceNumber: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  nationality: string;
  dateOfBirth: string;
  passportNumber: string;
  passportExpiry: string;
  currentAddress: string;
  permitType: string;
  employerName: string;
  employerCountry: string;
  jobTitle: string;
  jobSalary: string;
  startDate: string;
  contractDuration: string;
  hasPassport: boolean;
  hasJobOffer: boolean;
  hasMedicalCert: boolean;
  hasCriminalRecord: boolean;
  hasPhotos: boolean;
  hasEducationCert: boolean;
  status: string;
  createdAt: string;
}

export interface UploadResponse {
  url: string;
  filename: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
}

export interface ContactInquiry {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface SiteStats {
  totalJobs: number;
  totalApplications: number;
  totalCategories: number;
  countriesServed: number;
}

export type ListJobsParams = {
  category?: string;
  location?: string;
};